try adjusting bboxes based on char length
